+++ types/components/utils/UluConditionalWrapper.vue.d.ts
@@ -1,0 +1,20 @@
+declare const _default: __VLS_WithSlots<typeof __VLS_component, __VLS_Slots>;
+export default _default;
+type __VLS_WithSlots<T, S> = T & (new () => {
+    $slots: S;
+});
+declare const __VLS_component: import("vue").DefineComponent<{}, {
+    $props: Partial<typeof __VLS_props>;
+    is: string | Function | Record<string, any>;
+    unwrapped: boolean;
+}, {}, {}, {}, import("vue").ComponentOptionsMixin, import("vue").ComponentOptionsMixin, {}, string, import("vue").PublicProps, Readonly<{}> & Readonly<{}>, {}, {}, {}, {}, string, import("vue").ComponentProvideOptions, true, {}, any>;
+type __VLS_Slots = {
+    default?: ((props: {}) => any) | undefined;
+} & {
+    default?: ((props: {}) => any) | undefined;
+};
+declare const __VLS_props: {
+    readonly is: string | Function | Record<string, any>;
+    readonly unwrapped: boolean;
+};
+//# sourceMappingURL=UluConditionalWrapper.vue.d.ts.map