+++ types/components/utils/UluRouteAnnouncer.vue.d.ts
@@ -1,7 +1,7 @@
 declare const _default: import("vue").DefineComponent<import("vue").ExtractPropTypes<{
     /**
      * Allow user to bypass this functionality
-     * - Function should return true if the page should be annouced
+     * - Function should return true if the page should be announced
      * - Function is passed  (to, from, $route) => {}
      *   - to/from are path strings
      */
@@ -30,7 +30,7 @@ declare const _default: import("vue").DefineComponent<import("vue").ExtractPropT
 }, {}, import("vue").ComponentOptionsMixin, import("vue").ComponentOptionsMixin, {}, string, import("vue").PublicProps, Readonly<import("vue").ExtractPropTypes<{
     /**
      * Allow user to bypass this functionality
-     * - Function should return true if the page should be annouced
+     * - Function should return true if the page should be announced
      * - Function is passed  (to, from, $route) => {}
      *   - to/from are path strings
      */