+++ types/components/utils/UluSanityRichText.vue.d.ts
@@ -1,0 +1,15 @@
+declare const _default: import("vue").DefineComponent<{}, {
+    $props: Partial<typeof __VLS_props>;
+    noWrapper: boolean;
+    content?: unknown[] | undefined;
+}, {}, {}, {}, import("vue").ComponentOptionsMixin, import("vue").ComponentOptionsMixin, {}, string, import("vue").PublicProps, Readonly<{}> & Readonly<{}>, {}, {}, {}, {}, string, import("vue").ComponentProvideOptions, true, {}, any>;
+export default _default;
+/**
+ * A component to render Sanity's Portable Text format.
+ * It requires the '@portabletext/vue' package to be installed by the consumer.
+ */
+declare const __VLS_props: {
+    readonly noWrapper: boolean;
+    readonly content?: unknown[] | undefined;
+};
+//# sourceMappingURL=UluSanityRichText.vue.d.ts.map